requirements.txt
altgraph==0.17.4
lxml==6.0.2
packaging==25.0
pillow==12.0.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
python-pptx==1.0.2
setuptools==80.9.0
typing_extensions==4.15.0
xlsxwriter==3.2.9
